requirements.txt
# Web框架
fastapi==0.116.1
uvicorn[standard]==0.30.6
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
alembic==1.13.2
sqlite3  # 内置模块

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# CORS
fastapi[all]==0.116.1

# 文件处理
python-docx==0.8.11
PyPDF2==3.0.1
openpyxl==3.1.2
aiofiles==23.2.1
xmind==1.2.0

# 数据处理
pandas==2.1.1

# 自然语言处理
jieba==0.42.1

# 验证
pydantic[email]==2.6.4
pydantic-settings==2.2.1

# 日志和监控
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 开发工具
black==23.12.1
isort==5.13.2
flake8==6.1.0